add ApiItemController

+++ src/main/java/myex/shopping/controller/api/ApiItemController.java
@@ -1,49 +1,47 @@
-package myex.shopping.controller;
+package myex.shopping.controller.api;
 
 import lombok.RequiredArgsConstructor;
 import myex.shopping.domain.Item;
 import myex.shopping.form.ItemAddForm;
 import myex.shopping.repository.ItemRepository;
 import myex.shopping.service.ItemService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.List;
-import java.util.UUID;
 
 
 @Controller
-@RequestMapping("/items")
+@RequestMapping("/api/items")
 @RequiredArgsConstructor
-public class ItemController {
+public class ApiItemController {
 
     private final ItemRepository itemRepository; //생성자 주입.
     private final ItemService itemService;
 
     //전체 아이템 조회
     @GetMapping
-    public String items(Model model) {
+    public ResponseEntity<List<Item>> items() {
         List<Item> items = itemRepository.findAll();
-        model.addAttribute("items", items);
-        return "items/items";
+        return ResponseEntity.ok(items);
     }
 
     //개별 아이템 상세 조회
     @GetMapping("/{itemId}")
-    public String item(@PathVariable long itemId, Model model) {
+    public ResponseEntity<Item> item(@PathVariable long itemId) {
         Item item = itemRepository.findById(itemId);
-        model.addAttribute("item", item);
-        return "items/item";
-
+        if (item != null)
+            return ResponseEntity.ok(item);
+        else
+            //바디 없이 상태 코드만 가진 ResponseEntity 반환 가능.
+            return ResponseEntity.notFound().build();
     }
-
+/*  Form 뷰를 호출 -> 프론트엔드.
     //아이템 추가
     @GetMapping("/add")
     public String addForm(Model model) {
@@ -51,13 +49,14 @@ public class ItemController {
         return "items/addForm";
     }
 
+    */
+
+    //아이템 추가 로직
+    //AddForm, 프론트에서 imageFile, itemName, price, quantity 넘어옴.
     @PostMapping("/add")
-    public String addItem(@ModelAttribute("item")ItemAddForm form,
+    public ResponseEntity<Item> addItem(@ModelAttribute("item")ItemAddForm form,
                           RedirectAttributes redirectAttributes) throws IOException {
 
-        //UploadFolder 에 있는 사진 업로드 시도 하면, "같은 경로 + 같은 파일명" 이라 같다고 판단해 move 불가능.(오류 발생. /UUID로 바꿀시, "같은 경로 + 다른 파일명" 이라 다른 파일이라 판단하고 업로드 가능.
-
-        //"다른 경로 + 같은 파일명" : 덮어쓰기 해버림.
         Item item = itemService.ImageSave(form, new Item());
         item.setItemName(form.getItemName());
         item.setPrice(form.getPrice());
@@ -70,20 +69,22 @@ public class ItemController {
         System.out.println(savedItem.getImageUrl());
         System.out.println(item.getImageUrl());
         System.out.println("ItemController.addItem : postmapping");;
-        return "redirect:/items/{itemId}";
+        return ResponseEntity.status(HttpStatus.CREATED).body(savedItem);
     }
 
-    //아이템 수정
-    @GetMapping("/{itemId}/edit")
-    public String editForm (@PathVariable("itemId") Long itemId,
-                            Model model) {
-        Item findItem = itemRepository.findById(itemId);
-        model.addAttribute("item", findItem);
-        return "items/editForm"; //이건 뿌리는 뷰 (서버 html 위치)
-    }
+    //아이템 수정 :상세 + 수정화면 공용
+    //수정 화면이 필요하면 GET /items/{itemId} 에서 현재 상태를 가져와서 수정 폼에 채움.
 
-    @PostMapping("/{itemId}/edit") //이게 URL
-    public String edit (@PathVariable Long itemId,
+    /*
+    보통 이렇게 설계.
+    GET /items -> 전체 목록 조회
+    GET /items/{itemId} -> 단일 아이템 조회 
+    PUT /items/{itemId} -> 수정 요청
+     */
+    
+    //한개만 수정 : PutMapping
+    @PutMapping("/{itemId}/edit") //이게 URL
+    public ResponseEntity<Item> editItem (@PathVariable Long itemId,
                         @ModelAttribute("item")ItemAddForm form) throws IOException {
 
 
@@ -93,21 +94,21 @@ public class ItemController {
         item.setQuantity(form.getQuantity());
 
         itemRepository.update(itemId, item);
-//        itemRepository.update_exceptImgUrl(itemId, item);
-        return "redirect:/items/{itemId}";
-        //  {} 치환 순위.
-        /*
-          1. RedirectAttributes.addAttribute("itemId",...)
-          2. @PathVariable, @RequestParma, 같은 요청
-             메서드 파라미터 이름 Long itemId 랑 매칭됨.
-        */
+        return ResponseEntity.ok(item);
     }
 
-    //아이템 삭제하기 넣기.
-    @PostMapping("/{itemId}/delete")
-    public String deleteItem(@PathVariable Long itemId) {
-        itemRepository.deleteItem(itemId);
-        return "redirect:/items";
+
+    @DeleteMapping("/{itemId}/delete")
+    public ResponseEntity<?> deleteItem(@PathVariable Long itemId) {
+
+        Item findItem = itemRepository.findById(itemId);
+        if (findItem != null) {
+            itemRepository.deleteItem(itemId);
+            return ResponseEntity.noContent().build(); //204 No Content
+        }
+        else {
+            return ResponseEntity.notFound().build(); //404 Not Found
+        }
     }
 
 }
